Refactor StubIdpConfigurationTest: use base test
This will make it easier to reuse the logic in other configuration tests

solo @tlwr
