+++ stub-idp/src/test/java/uk/gov/ida/stub/idp/configuration/StubIdpConfigurationTest.java
@@ -1,20 +1,15 @@
 package uk.gov.ida.stub.idp.configuration;
 
-import io.dropwizard.configuration.ConfigurationSourceProvider;
 import io.dropwizard.configuration.YamlConfigurationFactory;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-
 import static io.dropwizard.jackson.Jackson.newObjectMapper;
 import static io.dropwizard.jersey.validation.Validators.newValidator;
 import static org.hamcrest.core.StringContains.containsString;
 
-public class StubIdpConfigurationTest {
+public class StubIdpConfigurationTest extends BaseConfigurationTest {
 
     @Rule
     public final ExpectedException expectedException = ExpectedException.none();
@@ -35,18 +30,4 @@ public class StubIdpConfigurationTest {
 
         factory.build(new StringConfigurationSourceProvider("url: "), "");
     }
-
-    class StringConfigurationSourceProvider implements ConfigurationSourceProvider {
-        private String configuration;
-
-        public StringConfigurationSourceProvider(String configuration) {
-            this.configuration = configuration;
-        }
-
-        @Override
-        public InputStream open(String path) throws IOException {
-            return new ByteArrayInputStream(this.configuration.getBytes());
-        }
-    }
-
 }
